misc!: remove node visitor feature
This feature was a relic of the first release of the library. It had
strong design issues and was going to become a huge blocker for upcoming
features.

Although it was never documented, it may have been used in applications;
there will be no replacement for this feature. If this becomes an issue
for existing applications, an issue can be created in the repository to
discuss other possible solutions.