misc: do not use `uniqid()`
The return value of `uniqid()` is not very random and easily guessable. While
this likely does not introduce any issues in CompiledPhpFileCache, the fact
that `uniqid()` is often mis-used in locations where actual randomness is a
hard requirement, makes this a red flag.

Replace the use of `uniqid()` to generate a random filename by hexadecimal
encoded `random_bytes(16)`, giving 128 Bits of randomness, making the value
unguessable in practice and preventing the `CompiledPhpFileCache` from showing
up when searching for `uniqid()`.