Remove @PHP8.0 annotations
Context:

As currently stands for the 1.14.x branch of the `doctrine/annotations`, code units with `@PHP8` annotations fail to be loaded.

This happens because the default list of implicitly ignored annotations does not contain `PHP8` in it, and therefore in scenarios where applications are loaded with components with such annotations fail with errors of this nature:

```sh
PHP Fatal error: Uncaught Doctrine\Common\Annotations\AnnotationException: [Semantical Error] The annotation "@PHP8" in method CuyZ\Valinor\Cache\FileSystemCache::get() was never imported. Did you maybe forget to add a "use" statement for this annotation? in vendor/doctrine/annotations/lib/Doctrine/Common/Annotations/AnnotationException.php:40
```

This change removes the annotation `@PHP8.0` and makes it a simple comment, which helps avoid applications running into issues because of the annotation not being loaded.