misc!: remove identifier attribute
This feature has been part of the library since its first public release, but it
was never documented because it did not fit one of the library's main philosophy
which is to be almost entirely decoupled from an application's domain layer.

The feature is entirely removed and not planned to be replaced by an
alternative, unless the community really feels like there is a need for
something alike.