fix: remove exception inheritance from `UnresolvableType`
Because `UnresolvableType` can be applied to any property, parameter or
function/method return type, it can be present in a class definition
hierarchy. This class definition can then be cached for better
application performance, meaning these `UnresolvableType` will be
present in the cache entry. This is a problem because an exception
contains a lot of information, including reflection instances which can
not be serialized and can break caching mechanisms that rely on
serialization.

`UnresolvableType` extending `LogicException` was not necessary, so it
is removed and the code relying on it has been adapted.