fix: properly handle class/enum name in shaped array key
This commit changes a bit the behaviour of the parser, which is now
unable to properly detect missing colon for class constants or enum
cases filters — for instance, in `SomeClass:SOME_CONSTANT`, the missing
colon was previously detected, now it won't be.

This is intentional, because this behaviour would limit the parser's
ability to properly parse shaped array keys containing class/enum name.

Because this was not such an important feature of the parser, there will
be no replacement for it and instead will fix the issue described above.