fix: improve cache warmup by creating required directories
Calling the mapper warmup will now go through all caches layers and run
a warmup process. When using `FileSystemCache` that is provided by the
library, the directories required by this cache will be created.

This should fix an issue that could occur when race conditions happened
in an app with a lot of traffic.