fix: load attributes lazily during runtime and cache access
Attributes are now loaded in a lazy way, preventing an error from
happening when attributes were compiled in the cache, but not present
when using the mapper during the runtime.

This could happen for instance with attributes used to generate
documentation of an API: they are loaded as a development dependency,
then the cache is warmed up during deployment, so they are present in
the mapper compiled cache; but these attributes are missing from the
live version of the application, so they cannot be accessed.

This commit helps to prevent a fatal error from happening in this kind
of situation.