misc: reorganize type resolver services
This is a major refactor of the classes used to resolve types from all
places: `@var`, `@param`, `@return`, `@extends`, `@template`,
`@phpstan-type`, `@phpstan-import-type`.

The goal was to remove remaining usages of regex to resolve the types,
and to move all parsing logic in dedicated classes instead of god
classes like the (now deleted) `ReflectionTypeResolver` and `DocParser`
services.