feat: introduce a benchmark system
This commit introduce a benchmark system based on phpbench.

In essence for each new pull request it runs benchmarks on the master branch and then
run the same benchmarks against the code from the pull request and compare them.

If the new code average execution time goes above 10% it will fail the pull request.

This commit only focuses into introducing this system and not create an exhaustive
set of benchmarks. There are still a few benchmarks created in order to verify the
system works.

It also adds a bit of documentation since the phpbench output can be hard to
understand if not already familiar with it.