qa: introduce a benchmark system
This commit introduces a benchmark system based on PHPBench.

This commit only focuses on introducing this system and not create an
exhaustive set of benchmarks. There are still a few benchmarks created
in order to verify the system works.

It also adds a bit of documentation since the PHPBench output can be
hard to understand if not already familiar with it.