feat: introduce compiled normalizer cache
This commit introduces a new compilation step for the normalizer, which
aims to bring huge performance gains when normalizing any value. It
works by adding a static analysis step to the process, which will
recursively analyse how the normalizer should perform for every value it
can meet. This process results in a native PHP code entry, that can
then be cached for further usage.

The PHP code is an anonymous class that implements the `Transformer`
interface, and is responsible of recursively transforming an input to
the same structure, but with only arrays and scalar values. It is done
by adding methods to the class, where basically each method is used to
transform a specific type.

Example #1: a method that transforms a string, with a registered custom
transformer.

```php
(new MapperBuilder())
    ->registerTransformer(fn (string $value) => strtoupper($value))
    ->normalizer(Format::array())
    ->normalize('some string');

// Generated method for a string value:
private function transform_string_xyz(mixed $value): mixed
{
    $next = fn () => $value;
    if (\is_string($value)) {
        $next = fn () => $this->transformers[0]($value, $next);
    }
    return $next();
}
```

Example #2: a method that transforms a property of a class, which is
typed as a shaped array.

```php
final readonly class SomeClass
{
    public function __construct(
        /** @var array{foo: string, bar: int} */
        public array $someArray,
    ) {}
}

// Generated method for the shaped array property:
private function transform_shaped_array_xyz(array $value): array
{
    $result = [];
    foreach ($value as $key => $item) {
        $result[$key] = match ($key) {
            'foo' => $this->transform_string_xyz($item),
            'bar' => $this->transform_int_xyz($item),
            default => $this->transform_mixed($item),
        };
    }
    return $result;
}
```

The entry point to this process can be found in `CacheTransformer` where
the following steps occur:

1. The value to be normalized is analyzed: an instance of `Type` is
   created, which will be the root type for the whole static analysis
   process.
2. The `TransformerDefinitionBuilder` will recursively fetch all types
   that can be met during the normalization. For instance, if the root
   type is a class, then all its properties and their types will be
   included in the process. For each type, transformers (registered with
   `MapperBuilder::registerTransformer`) are analyzed: if a transformer
   matches a type, it will be part of the compilation process.
3. For each type, a `TypeFormatter` will be called. These type
   formatters are responsible for manipulating the compiler which is
   used to create the PHP code that will normalize the value. How it is
   done really depends on the type and can be really simple (e.g.
   `DateTimeFormatter`) or quite complex (e.g. `ClassFormatter`).
4. The PHP code is then written into a file and included when needed.