misc!: introduce internal cache interface and remove PSR-16 dependency
The Simple Cache PSR requirement has been removed from the library. This
means that external PSR implementations can no longer be used with this
library.

This decision was taken because entries that were cached were always
generated PHP code, which is not what PSR is meant for. This led to
confusion and incorrect behavior when using the library with a custom
implementation.

The library now uses its own cache interface, but the provided
implementations did not change, thus the upgrade should be pretty
straightforward.

Also note that this change led to a refactor that should improve
confidence for the file watching feature: before, the cache entries were
sometimes not invalidated properly when files changed during
development. This should now be better.