feat!: introduce `NormalizerBuilder` as the main entry for normalizers
The `NormalizerBuilder` class has been introduced and will now be the
main entry to instantiate normalizers. Therefore, the methods in
`MapperBuilder` that used to configure and return normalizers have been
removed.

This decision aims to make a clear distinction between the mapper and
the normalizer configuration API, where confusion could arise when using
both.

The `NormalizerBuilder` can be used like this:

```php
$normalizer = (new \CuyZ\Valinor\NormalizerBuilder())
    ->registerTransformer(
        fn (\DateTimeInterface $date) => $date->format('Y/m/d')
    )
    ->normalizer(\CuyZ\Valinor\Normalizer\Format::array())
    ->normalize($someData);
```