feat: Add PHPUnit extension for pretty-printing MappingError instances
This helps with debugging MappingErrors thrown during test execution.
Sometimes these can be annoying to debug because Valinor only shows "A
total of 2 errors were thrown".

The extension works by overriding the TestCase `transformException`
method through a trait, which means this is an opt-in feature for tests.
It collects them in a singleton class and then uses the PHPUnit
extension to display them in a formatted table at the end of the test
run.

Implementing it is a matter of using the `CollectValinorMappingErrors`
trait in your TestCase class and registering the
`PrettyPrintMappingErrorsExtension` extension class in the phpunit
configuration file. A PR to update the documentation will be made
separately.

This does require symfony/console for its table printing capabilities,
copying that logic into the library seemed excessive. In order to not
have a hard dependency on it, it is put in the suggest array in
composer.json and we are using a class_exists check to see if it is
installed.

Developed by: Social Deal (@socialdeal)

Signed-off-by: NanoSector <rick@nanosector.nl>