feat: improve mapping error messages types signatures
This greatly improves the feedback given in mapping errors, especially
the expected signature of the current node.

This gets rid of the infamous `?` that was used whenever an object was
present in a type, leading to incomplete and misleading messages.

Example of a new message:

```php
final class User
{
    public function __construct(
        public string $name,
        public int $age,
    ) {}
}

(new MapperBuilder())
    ->mapper()
    ->map(User::class, 'invalid value');

// Could not map type `User`. An error occurred at path *root*: Value
// 'invalid value' does not match `array{name: string, age: int}`.
```

Co-authored-by: Romain Canon <romain.hydrocanon@gmail.com>