chore: refactor class and interface mapping process
This refactor aims to ease the upcoming maintenance of the library, by
making the class and interface node builders process smoother.

The main change is that these builders now share the same logic
concerning the children nodes building. Instead of having their own
logic, they convert the needed structures signatures to a shaped array,
that is then given back to the mapping process. This way, this complex
logic is now located in one place instead of three, allowing easier
maintenance as well as more capabilities.